Update color module with new direction and interpolation functions.
- Added functions for getting direction and interpolation.
- Removed duplicate utility functions.
